DA-833 Fix NRRD image distortion
The problem is the spacing used in (ijk2LPS) core.utils.js multiplies spacing y before spacing x.
This probably works with DICOM images (as mentioned in the comments on the code by original developers)
But it doesn't work with NRRD.
The way I resolved the issue was by tracing the code from model loading and where ijk2LPS calculations happen
and I figured out changing spacing vector resolves the issue.
This should solve the issue for now.